Move the feed registry from Feeds to the config
