+++ lib/solidus_feeds.rb
@@ -5,6 +5,4 @@ require 'solidus_support'
 
 require 'solidus_feeds/configuration'
 require 'solidus_feeds/version'
-require 'solidus_feeds/feed'
-require 'solidus_feeds/feeds'
 require 'solidus_feeds/engine'
